Update SendEmailTask to make it work with Solicitors and Citizen
