+++ src/main/java/uk/gov/hmcts/reform/divorce/orchestration/tasks/servicejourney/emails/DeemedApprovedEmailTask.java
@@ -36,7 +36,7 @@ public class DeemedApprovedEmailTask extends SendEmailTask {
     }
 
     @Override
-    protected EmailTemplateNames getTemplate() {
+    protected EmailTemplateNames getTemplate(Map<String, Object> caseData) {
         return EmailTemplateNames.CITIZEN_DEEMED_APPROVED;
     }
 }
