+++ src/main/java/uk/gov/hmcts/reform/divorce/orchestration/tasks/servicejourney/emails/DispensedApprovedEmailTask.java
@@ -10,14 +10,19 @@ import uk.gov.hmcts.reform.divorce.orchestration.service.EmailService;
 
 import java.util.Map;
 
+import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.OrchestrationConstants.NOTIFICATION_CCD_REFERENCE_KEY;
 import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.OrchestrationConstants.NOTIFICATION_PET_NAME;
+import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.OrchestrationConstants.NOTIFICATION_RESP_NAME;
+import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.OrchestrationConstants.NOTIFICATION_SOLICITOR_NAME;
 import static uk.gov.hmcts.reform.divorce.orchestration.service.bulk.print.dataextractor.FullNamesDataExtractor.getPetitionerFullName;
+import static uk.gov.hmcts.reform.divorce.orchestration.service.bulk.print.dataextractor.FullNamesDataExtractor.getPetitionerSolicitorFullName;
+import static uk.gov.hmcts.reform.divorce.orchestration.service.bulk.print.dataextractor.FullNamesDataExtractor.getRespondentFullName;
+import static uk.gov.hmcts.reform.divorce.orchestration.tasks.util.TaskUtils.getCaseId;
+import static uk.gov.hmcts.reform.divorce.orchestration.util.PartyRepresentationChecker.isPetitionerRepresented;
 
 @Component
 @Slf4j
 public class DispensedApprovedEmailTask extends SendEmailTask {
-
-    private static final String SUBJECT = "Your ‘dispense with service’ application has been approved";
 
     public DispensedApprovedEmailTask(EmailService emailService) {
         super(emailService);
@@ -25,18 +30,35 @@ public class DispensedApprovedEmailTask extends SendEmailTask {
 
     @Override
     protected String getSubject(Map<String, Object> caseData) {
-        return SUBJECT;
+        if (isPetitionerRepresented(caseData)) {
+            return "%s vs %s: Dispense with service application has been approved";
+        } else {
+            return "Your ‘dispense with service’ application has been approved";
+        }
     }
 
     @Override
     protected Map<String, String> getPersonalisation(TaskContext taskContext, Map<String, Object> caseData) {
-        return ImmutableMap.of(
-            NOTIFICATION_PET_NAME, getPetitionerFullName(caseData)
-        );
+        if (isPetitionerRepresented(caseData)) {
+            return ImmutableMap.of(
+                NOTIFICATION_PET_NAME, getPetitionerFullName(caseData),
+                NOTIFICATION_RESP_NAME, getRespondentFullName(caseData),
+                NOTIFICATION_CCD_REFERENCE_KEY, getCaseId(taskContext),
+                NOTIFICATION_SOLICITOR_NAME, getPetitionerSolicitorFullName(caseData)
+            );
+        } else {
+            return ImmutableMap.of(
+                NOTIFICATION_PET_NAME, getPetitionerFullName(caseData)
+            );
+        }
     }
 
     @Override
-    protected EmailTemplateNames getTemplate() {
-        return EmailTemplateNames.CITIZEN_DISPENSED_APPROVED;
+    protected EmailTemplateNames getTemplate(Map<String, Object> caseData) {
+        if (isPetitionerRepresented(caseData)) {
+            return EmailTemplateNames.SOL_DISPENSED_APPROVED;
+        } else {
+            return EmailTemplateNames.CITIZEN_DISPENSED_APPROVED;
+        }
     }
 }
