+++ src/main/java/uk/gov/hmcts/reform/divorce/orchestration/workflows/servicejourney/ServiceDecisionMadeWorkflow.java
@@ -13,7 +13,6 @@ import uk.gov.hmcts.reform.divorce.orchestration.tasks.servicejourney.emails.Dee
 import uk.gov.hmcts.reform.divorce.orchestration.tasks.servicejourney.emails.DispensedApprovedEmailTask;
 import uk.gov.hmcts.reform.divorce.orchestration.tasks.servicejourney.emails.DispensedNotApprovedEmailTask;
 import uk.gov.hmcts.reform.divorce.orchestration.tasks.servicejourney.emails.SolicitorDeemedApprovedEmailTask;
-import uk.gov.hmcts.reform.divorce.orchestration.tasks.servicejourney.emails.SolicitorDispensedApprovedEmailTask;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -37,7 +36,6 @@ public class ServiceDecisionMadeWorkflow extends DefaultWorkflow<Map<String, Obj
     private final DispensedApprovedEmailTask dispensedApprovedEmailTask;
     private final DispensedNotApprovedEmailTask dispensedNotApprovedEmailTask;
     private final SolicitorDeemedApprovedEmailTask solicitorDeemedApprovedEmailTask;
-    private final SolicitorDispensedApprovedEmailTask solicitorDispensedApprovedEmailTask;
 
     public Map<String, Object> run(CaseDetails caseDetails, String authorisation)
         throws WorkflowException {
@@ -91,16 +89,7 @@ public class ServiceDecisionMadeWorkflow extends DefaultWorkflow<Map<String, Obj
     }
 
     private void sendDispensedApprovedEmail(Map<String, Object> caseData, String caseId, List<Task<Map<String, Object>>> tasks) {
-        if (isPetitionerRepresented(caseData)) {
-            tasks.add(getTaskForSolicitorDispensedApproved(caseId));
-        } else {
-            tasks.add(getTaskForDispensedApproved(caseId));
-        }
-    }
-
-    private Task<Map<String, Object>> getTaskForSolicitorDispensedApproved(String caseId) {
-        log.info("CaseId: {} dispensed solicitor email task adding.", caseId);
-        return solicitorDispensedApprovedEmailTask;
+        tasks.add(getTaskForDispensedApproved(caseId));
     }
 
     private Task<Map<String, Object>> getTaskForDispensedApproved(String caseId) {
