--- src/test/java/uk/gov/hmcts/reform/divorce/orchestration/tasks/servicejourney/emails/SolicitorDispensedApprovedEmailTaskTest.java
@@ -1,104 +1,0 @@
-package uk.gov.hmcts.reform.divorce.orchestration.tasks.servicejourney.emails;
-
-import com.google.common.collect.ImmutableMap;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.mockito.InjectMocks;
-import org.mockito.Mock;
-import org.mockito.junit.MockitoJUnitRunner;
-import uk.gov.hmcts.reform.divorce.orchestration.domain.model.LanguagePreference;
-import uk.gov.hmcts.reform.divorce.orchestration.framework.workflow.task.DefaultTaskContext;
-import uk.gov.hmcts.reform.divorce.orchestration.framework.workflow.task.TaskContext;
-import uk.gov.hmcts.reform.divorce.orchestration.framework.workflow.task.TaskException;
-import uk.gov.hmcts.reform.divorce.orchestration.service.EmailService;
-
-import java.util.HashMap;
-import java.util.Map;
-
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.equalTo;
-import static org.mockito.Mockito.verify;
-import static uk.gov.hmcts.reform.divorce.orchestration.TestConstants.TEST_CASE_ID;
-import static uk.gov.hmcts.reform.divorce.orchestration.TestConstants.TEST_PETITIONER_FIRST_NAME;
-import static uk.gov.hmcts.reform.divorce.orchestration.TestConstants.TEST_PETITIONER_FULL_NAME;
-import static uk.gov.hmcts.reform.divorce.orchestration.TestConstants.TEST_PETITIONER_LAST_NAME;
-import static uk.gov.hmcts.reform.divorce.orchestration.TestConstants.TEST_RESPONDENT_FIRST_NAME;
-import static uk.gov.hmcts.reform.divorce.orchestration.TestConstants.TEST_RESPONDENT_FULL_NAME;
-import static uk.gov.hmcts.reform.divorce.orchestration.TestConstants.TEST_RESPONDENT_LAST_NAME;
-import static uk.gov.hmcts.reform.divorce.orchestration.TestConstants.TEST_SOLICITOR_EMAIL;
-import static uk.gov.hmcts.reform.divorce.orchestration.TestConstants.TEST_SOLICITOR_NAME;
-import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.OrchestrationConstants.CASE_ID_JSON_KEY;
-import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.OrchestrationConstants.D_8_PETITIONER_FIRST_NAME;
-import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.OrchestrationConstants.D_8_PETITIONER_LAST_NAME;
-import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.OrchestrationConstants.NOTIFICATION_CCD_REFERENCE_KEY;
-import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.OrchestrationConstants.NOTIFICATION_PET_NAME;
-import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.OrchestrationConstants.NOTIFICATION_RESP_NAME;
-import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.OrchestrationConstants.NOTIFICATION_SOLICITOR_NAME;
-import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.OrchestrationConstants.PETITIONER_SOLICITOR_EMAIL;
-import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.OrchestrationConstants.PETITIONER_SOLICITOR_NAME;
-import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.OrchestrationConstants.RESP_FIRST_NAME_CCD_FIELD;
-import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.OrchestrationConstants.RESP_LAST_NAME_CCD_FIELD;
-import static uk.gov.hmcts.reform.divorce.orchestration.domain.model.email.EmailTemplateNames.SOL_DISPENSED_APPROVED;
-
-@RunWith(MockitoJUnitRunner.class)
-public class SolicitorDispensedApprovedEmailTaskTest {
-
-    @Mock
-    private EmailService emailService;
-
-    @InjectMocks
-    private SolicitorDispensedApprovedEmailTask task;
-
-    private static Map<String, String> getExpectedNotificationTemplateVars() {
-        return ImmutableMap.of(
-            NOTIFICATION_PET_NAME, TEST_PETITIONER_FULL_NAME,
-            NOTIFICATION_RESP_NAME, TEST_RESPONDENT_FULL_NAME,
-            NOTIFICATION_CCD_REFERENCE_KEY, TEST_CASE_ID,
-            NOTIFICATION_SOLICITOR_NAME, TEST_SOLICITOR_NAME
-        );
-    }
-
-    private static Map<String, Object> buildCaseData() {
-        Map caseData = new HashMap<>();
-        caseData.put(D_8_PETITIONER_FIRST_NAME, TEST_PETITIONER_FIRST_NAME);
-        caseData.put(D_8_PETITIONER_LAST_NAME, TEST_PETITIONER_LAST_NAME);
-        caseData.put(RESP_FIRST_NAME_CCD_FIELD, TEST_RESPONDENT_FIRST_NAME);
-        caseData.put(RESP_LAST_NAME_CCD_FIELD, TEST_RESPONDENT_LAST_NAME);
-        caseData.put(PETITIONER_SOLICITOR_EMAIL, TEST_SOLICITOR_EMAIL);
-        caseData.put(PETITIONER_SOLICITOR_NAME, TEST_SOLICITOR_NAME);
-        return caseData;
-    }
-
-    private static TaskContext getTaskContext() {
-        TaskContext context = new DefaultTaskContext();
-        context.setTransientObject(CASE_ID_JSON_KEY, TEST_CASE_ID);
-
-        return context;
-    }
-
-    private void verifyEmailSentSuccessfully(Map<String, Object> caseData) {
-        verify(emailService).sendEmail(
-            TEST_SOLICITOR_EMAIL,
-            SOL_DISPENSED_APPROVED.name(),
-            getExpectedNotificationTemplateVars(),
-            task.getSubject(caseData),
-            LanguagePreference.ENGLISH
-        );
-    }
-
-
-    private void executeTask() throws TaskException {
-        Map<String, Object> returnedPayload = task.execute(getTaskContext(), buildCaseData());
-
-        assertThat(returnedPayload, equalTo(buildCaseData()));
-    }
-
-    @Test
-    public void whenExecuteEmailNotificationTask_thenSendEmail() {
-        Map<String, Object> caseData = buildCaseData();
-
-        executeTask();
-
-        verifyEmailSentSuccessfully(caseData);
-    }
-}
